+++ functions.php
@@ -950,7 +950,7 @@ class OCEANWP_Theme_Class {
 	public static function meta_tags() {
 
 		// Return if disabled or if Yoast SEO enabled as they have their own meta tags
-		if ( false == get_theme_mod( 'ocean_opengraph', true )
+		if ( false == get_theme_mod( 'ocean_open_graph', false )
 			|| defined( 'WPSEO_VERSION' ) ) {
 			return;
 		}