+++ woocommerce/owp-single-product.php
@@ -38,8 +38,7 @@ foreach ( $elements as $element ) {
 	}
 
 	// Quantity & Add to cart button
-	if ( 'quantity-button' == $element
-		|| 'quantity_button' == $element ) {
+	if ( 'quantity-button' == $element ) {
 		woocommerce_template_single_add_to_cart();
 	}